Refactor the code, move the validation inside service to be applied when call from local service via rest-communication
