+++ src/core/test/base-service.utils/base-service.utils.all.tests.js
@@ -3,9 +3,10 @@
  *
  * include them in order
  */
+// require('./base-service.utils.getAllForReq.test.js');
 require('./base-service.utils.getAll.test.js');
-require('./base-service.utils.getOne.test.js');
-require('./base-service.utils.post.test.js');
-require('./base-service.utils.delete.test.js');
-require('./base-service.utils.put.test.js');
-require('./base-service.utils.patch.test.js');
+// require('./base-service.utils.getOne.test.js');
+// require('./base-service.utils.post.test.js');
+// require('./base-service.utils.delete.test.js');
+// require('./base-service.utils.put.test.js');
+// require('./base-service.utils.patch.test.js');
