+++ src/core/test/core.all.tests.js
@@ -4,7 +4,9 @@
  * include test files in order
  */
 require('./common.utils/common.utils.all.tests.js');
+
 require('./rest-api.utils/rest-api.utils.all.tests.js');
 require('./rest-messages.utils/rest-messages.utils.all.tests.js');
-require('./base-service.utils/base-service.utils.all.tests.js');
 require('./rest-communications.utils/rest-communications.utils.all.tests.js');
+
+require('./base-service.utils/base-service.utils.all.tests.js');
