+++ src/services/config.services.js
@@ -30,11 +30,6 @@ const Public = {
       await database.init();
     }
 
-    // init services
-    for (const service of [SchoolsService, UsersService, StudentsService]) {
-      await service.init();
-    }
-
     // init the communication
     const config = {
       local: {
