+++ src/services/users/users.database.js
@@ -1,0 +1,21 @@
+/**
+ * Database
+ */
+
+const Public = {
+  /**
+   * init
+   */
+  init: async () => {
+    console.log('Init users database');
+  },
+
+  /**
+   * get collection
+   */
+  collection: async (_ctx) => {
+    return null;
+  },
+};
+
+module.exports = { ...Public };
