+++ src/test/tests.js
@@ -6,6 +6,6 @@
 require('../core/test/core.all.tests.js');
 
 // services tests
-require('./test/schools/schools.all.tests.js');
+// require('./test/schools/schools.all.tests.js');
 // require('./test/users/users.all.tests.js');
 // require('./test/students/students.all.tests.js');
